move color components to common module
